added adc, fsm functionality
too many changes to list

+++ pypilot_eeprom.h
@@ -1,13 +1,13 @@
 #ifndef _PYPILOT_EEPROM_H_
 #define _PYPILOT_EEPROM_H_
 
-#include <inttypes.h>
+#include <stdint.h>
 
 /*-----------------------------------------------------------------------*/
 
-extern uint8_t eeprom_read8(uint16_t address);
+extern uint8_t eeprom_read8(uint8_t address);
 
-extern void eeprom_update8(uint16_t address, uint8_t val);
+extern void eeprom_update8(uint8_t address, uint8_t val);
 
 /*-----------------------------------------------------------------------*/
